实现高德服务的添加，设备的添加功能

+++ src/main/java/com/yibo/parking/service/Impl/car/DeviceServiceImpl.java
@@ -1,11 +1,13 @@
 package com.yibo.parking.service.Impl.car;
 
+import com.alibaba.fastjson.JSONObject;
 import com.yibo.parking.dao.car.DeviceMapper;
 import com.yibo.parking.dao.system.SystemDataMapper;
-import com.yibo.parking.dao.system.SystemServerMapper;
 import com.yibo.parking.entity.car.Device;
+import com.yibo.parking.entity.system.SystemData;
 import com.yibo.parking.service.DeviceService;
 import com.yibo.parking.utils.EntityIdGenerate;
+import com.yibo.parking.utils.HttpClientUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,9 +24,6 @@ public class DeviceServiceImpl implements DeviceService {
     @Autowired
     private SystemDataMapper dataMapper;
 
-    @Autowired
-    private SystemServerMapper serverMapper;
-
     @Override
     public List<Device> findList(Device device) {
         return deviceMapper.findList(device);
@@ -38,18 +37,59 @@ public class DeviceServiceImpl implements DeviceService {
     @Override
     public int save(Device device) {
         if (device.getId() != null){
+            return update(device);
+        }
+        return insert(device);
+    }
+
+    private int update(Device device){
+        SystemData data = dataMapper.getByKey("高德", "key");
+        String url = "https://tsapi.amap.com/v1/track/terminal/update";
+        Map<String, String> map = new HashMap<>();
+        map.put("key",data.getValue());
+        map.put("sid",device.getsId());
+        map.put("tid",device.getDeviceId());
+        String result = HttpClientUtil.doPost(url,map);
+        JSONObject object = JSONObject.parseObject(result);
+        if (object.get("errcode").toString().equals("10000")){
             return deviceMapper.update(device);
         }
-//        String key = dataMapper.getByKey("高德","key");
-//        SystemServer server = serverMapper.findGDByKey();
+        return 0;
+    }
+
+    private int insert(Device device){
         device.setId(EntityIdGenerate.generateId());
         device.setStatus("0");
-        return deviceMapper.insert(device);
+        SystemData data = dataMapper.getByKey("高德","key");
+        String url = "https://tsapi.amap.com/v1/track/terminal/add";
+        Map<String,String> map = new HashMap<>();
+        map.put("key",data.getValue());
+        map.put("sid",device.getsId());
+        map.put("name",device.getDeviceId());
+        String result = HttpClientUtil.doPost(url,map);
+        JSONObject object = JSONObject.parseObject(result);
+        if (object.get("errcode").toString().equals("10000")){
+            JSONObject json = object.getJSONObject("data");
+            device.settId(json.get("tid").toString());
+            return deviceMapper.insert(device);
+        }
+        return 0;
     }
 
     @Override
     public int delete(Device device) {
-        return deviceMapper.delete(device);
+        SystemData data = dataMapper.getByKey("高德","key");
+        String url = "https://tsapi.amap.com/v1/track/terminal/delete";
+        Map<String,String> map = new HashMap<>();
+        map.put("key",data.getValue());
+        map.put("sid",device.getsId());
+        map.put("tid",device.getDeviceId());
+        String result = HttpClientUtil.doPost(url,map);
+        JSONObject object = JSONObject.parseObject(result);
+        if (object.get("errcode").toString().equals("10000")){
+            return deviceMapper.delete(device);
+        }
+        return 0;
     }
 
     public Map<String,Object> status(Device device) {
